pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[project]
name = "local-pc-agent"
version = "0.1.0"
description = "Windows local PC control agent MVP (harness-first, no LLM)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyautogui>=0.9.54",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "Pillow>=10.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
